fix: AddBuff를 통한 Attack 데미지 전환, Get을 통한 정교화된 값을 통해 Attack의 공격속도 및 크기 조절

+++ Assets/Trieyes/Scripts/AttackSystem/AttackComponents/AC008_LightningField.cs
@@ -15,7 +15,7 @@ namespace AttackComponents
     public class AC008_LightningField : AttackComponent
     {
         [Header("번개 장판 설정")]
-        public float lightningFieldDamage = 30f;
+        public int lightningFieldDamage = 30;
         public float lightningFieldRadius = 2.5f;
         public float lightningFieldDuration = 3f;
         public float lightningFieldDelay = 0.1f;
